content redistribution

+++ src/peer_server/state.cpp
@@ -10,10 +10,10 @@ int total_content;
 
 std::unordered_map<uint8_t, char *> data;
 
-int ceiling() {
+uint16_t ceiling() {
     return (total_content + total_peers - 1) / total_peers;
 }
 
-int floor() {
+uint16_t floor() {
     return total_content / total_peers;
 }
